final resolve repo
