+++ br/pkg/conn/conn_test.go
@@ -9,11 +9,13 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/kvproto/pkg/metapb"
-	"github.com/pingcap/tidb/br/pkg/pdutil"
 	"github.com/stretchr/testify/require"
 	pd "github.com/tikv/pd/client"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+
+	"github.com/pingcap/tidb/br/pkg/conn/util"
+	"github.com/pingcap/tidb/br/pkg/pdutil"
 )
 
 type fakePDClient struct {
@@ -60,7 +62,7 @@ func TestGetAllTiKVStoresWithRetryCancel(t *testing.T) {
 		stores: stores,
 	}
 
-	_, err := GetAllTiKVStoresWithRetry(ctx, fpdc, SkipTiFlash)
+	_, err := GetAllTiKVStoresWithRetry(ctx, fpdc, util.SkipTiFlash)
 	require.Error(t, err)
 	require.Equal(t, codes.Canceled, status.Code(errors.Cause(err)))
 }
@@ -100,7 +102,7 @@ func TestGetAllTiKVStoresWithUnknown(t *testing.T) {
 		stores: stores,
 	}
 
-	_, err := GetAllTiKVStoresWithRetry(ctx, fpdc, SkipTiFlash)
+	_, err := GetAllTiKVStoresWithRetry(ctx, fpdc, util.SkipTiFlash)
 	require.Error(t, err)
 	require.Equal(t, codes.Unknown, status.Code(errors.Cause(err)))
 }
@@ -155,12 +157,12 @@ func TestCheckStoresAlive(t *testing.T) {
 		stores: stores,
 	}
 
-	kvStores, err := GetAllTiKVStoresWithRetry(ctx, fpdc, SkipTiFlash)
+	kvStores, err := GetAllTiKVStoresWithRetry(ctx, fpdc, util.SkipTiFlash)
 	require.NoError(t, err)
 	require.Len(t, kvStores, 2)
 	require.Equal(t, stores[2:], kvStores)
 
-	err = checkStoresAlive(ctx, fpdc, SkipTiFlash)
+	err = checkStoresAlive(ctx, fpdc, util.SkipTiFlash)
 	require.NoError(t, err)
 }
 
@@ -169,7 +171,7 @@ func TestGetAllTiKVStores(t *testing.T) {
 
 	testCases := []struct {
 		stores         []*metapb.Store
-		storeBehavior  StoreBehavior
+		storeBehavior  util.StoreBehavior
 		expectedStores map[uint64]int
 		expectedError  string
 	}{
@@ -177,22 +179,14 @@ func TestGetAllTiKVStores(t *testing.T) {
 			stores: []*metapb.Store{
 				{Id: 1},
 			},
-			storeBehavior:  SkipTiFlash,
+			storeBehavior:  util.SkipTiFlash,
 			expectedStores: map[uint64]int{1: 1},
 		},
 		{
 			stores: []*metapb.Store{
 				{Id: 1},
 			},
-			storeBehavior:  ErrorOnTiFlash,
-			expectedStores: map[uint64]int{1: 1},
-		},
-		{
-			stores: []*metapb.Store{
-				{Id: 1},
-				{Id: 2, Labels: []*metapb.StoreLabel{{Key: "engine", Value: "tiflash"}}},
-			},
-			storeBehavior:  SkipTiFlash,
+			storeBehavior:  util.ErrorOnTiFlash,
 			expectedStores: map[uint64]int{1: 1},
 		},
 		{
@@ -200,7 +194,15 @@ func TestGetAllTiKVStores(t *testing.T) {
 				{Id: 1},
 				{Id: 2, Labels: []*metapb.StoreLabel{{Key: "engine", Value: "tiflash"}}},
 			},
-			storeBehavior: ErrorOnTiFlash,
+			storeBehavior:  util.SkipTiFlash,
+			expectedStores: map[uint64]int{1: 1},
+		},
+		{
+			stores: []*metapb.Store{
+				{Id: 1},
+				{Id: 2, Labels: []*metapb.StoreLabel{{Key: "engine", Value: "tiflash"}}},
+			},
+			storeBehavior: util.ErrorOnTiFlash,
 			expectedError: "cannot restore to a cluster with active TiFlash stores.*",
 		},
 		{
@@ -212,7 +214,7 @@ func TestGetAllTiKVStores(t *testing.T) {
 				{Id: 5, Labels: []*metapb.StoreLabel{{Key: "else", Value: "tikv"}, {Key: "engine", Value: "tiflash"}}},
 				{Id: 6, Labels: []*metapb.StoreLabel{{Key: "else", Value: "tiflash"}, {Key: "engine", Value: "tikv"}}},
 			},
-			storeBehavior:  SkipTiFlash,
+			storeBehavior:  util.SkipTiFlash,
 			expectedStores: map[uint64]int{1: 1, 3: 1, 4: 1, 6: 1},
 		},
 		{
@@ -224,7 +226,7 @@ func TestGetAllTiKVStores(t *testing.T) {
 				{Id: 5, Labels: []*metapb.StoreLabel{{Key: "else", Value: "tikv"}, {Key: "engine", Value: "tiflash"}}},
 				{Id: 6, Labels: []*metapb.StoreLabel{{Key: "else", Value: "tiflash"}, {Key: "engine", Value: "tikv"}}},
 			},
-			storeBehavior: ErrorOnTiFlash,
+			storeBehavior: util.ErrorOnTiFlash,
 			expectedError: "cannot restore to a cluster with active TiFlash stores.*",
 		},
 		{
@@ -236,14 +238,14 @@ func TestGetAllTiKVStores(t *testing.T) {
 				{Id: 5, Labels: []*metapb.StoreLabel{{Key: "else", Value: "tikv"}, {Key: "engine", Value: "tiflash"}}},
 				{Id: 6, Labels: []*metapb.StoreLabel{{Key: "else", Value: "tiflash"}, {Key: "engine", Value: "tikv"}}},
 			},
-			storeBehavior:  TiFlashOnly,
+			storeBehavior:  util.TiFlashOnly,
 			expectedStores: map[uint64]int{2: 1, 5: 1},
 		},
 	}
 
 	for _, testCase := range testCases {
 		pdClient := fakePDClient{stores: testCase.stores}
-		stores, err := GetAllTiKVStores(context.Background(), pdClient, testCase.storeBehavior)
+		stores, err := util.GetAllTiKVStores(context.Background(), pdClient, testCase.storeBehavior)
 		if len(testCase.expectedError) != 0 {
 			require.Error(t, err)
 			require.Regexp(t, testCase.expectedError, err.Error())
