+++ br/pkg/restore/split/split_client.go
@@ -31,7 +31,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 
-	"github.com/pingcap/tidb/br/pkg/conn"
+	"github.com/pingcap/tidb/br/pkg/conn/util"
 	errors2 "github.com/pingcap/tidb/br/pkg/errors"
 	"github.com/pingcap/tidb/br/pkg/httputil"
 	"github.com/pingcap/tidb/br/pkg/logutil"
@@ -494,7 +494,7 @@ func (c *pdClient) BatchSplitRegions(
 }
 
 func (c *pdClient) getStoreCount(ctx context.Context) (int, error) {
-	stores, err := conn.GetAllTiKVStores(ctx, c.client, conn.SkipTiFlash)
+	stores, err := util.GetAllTiKVStores(ctx, c.client, util.SkipTiFlash)
 	if err != nil {
 		return 0, err
 	}
