+++ br/pkg/utils/utilpool/worker.go
@@ -1,6 +1,6 @@
 // Copyright 2020 PingCAP, Inc. Licensed under Apache-2.0.
 
-package utils
+package utilpool
 
 import (
 	"github.com/pingcap/log"
